feat : implement router and add pages
