+++ src/ChattingList.js
@@ -1,0 +1,8 @@
+import React from 'react';
+
+function ChattingList(){
+    return(
+        <div></div>
+    );
+}
+export default ChattingList;
