+++ src/FriendList.js
@@ -1,0 +1,8 @@
+import React from 'react';
+
+function FreindList(){
+    return(
+        <div></div>
+    );
+}
+export default FreindList;
