+++ src/SettingPage.js
@@ -1,0 +1,8 @@
+import React from 'react';
+
+function SettingPage(){
+    return(
+        <div></div>
+    );
+}
+export default SettingPage;
